+++ examples/check_s3_bucket.py
@@ -1,0 +1,8 @@
+import xrlint.all as xrl
+
+URL = "s3://xcube-test/"
+
+xrlint = xrl.XRLint(no_config_lookup=True)
+xrlint.init_config("recommended")
+results = xrlint.validate_files([URL])
+print(xrlint.format_results(results))